refactor: market event and ticket files
