+++ components/market/ticket/TicketCardList.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+function TicketCardList() {
+  return <div>TicketCardList</div>
+}
+
+export default TicketCardList
