--- components/ticket/TicketCard.tsx
@@ -1,7 +1,0 @@
-import React from "react"
-
-function TicketCard() {
-  return <div>TicketCard</div>
-}
-
-export default TicketCard
