+++ interfaces/event/event.interface.ts
@@ -3,9 +3,9 @@ import { PlaceType } from "./location.interface"
 export type TicketType = "GENERAL" | "VIP" | "RESERVED_SEAT"
 export type TicketTypeKey = "general" | "vip" | "reservedSeat"
 export const TicketTypeLabel = {
-  generalTickets: "General",
-  vipTickets: "VIP",
-  reservedSeatTickets: "ReservedSeat"
+  general: "General",
+  vip: "VIP",
+  reservedSeat: "ReservedSeat"
 }
 
 export type EventStatusType =
