+++ pages/marketplace/organizer/[organizerId].tsx
@@ -2,7 +2,7 @@ import { Container } from "@mui/material"
 import { Box } from "@mui/system"
 import { useRouter } from "next/router"
 import React, { useContext, useEffect } from "react"
-import OrganizerEventList from "../../../components/market/event/OrganizerEventList"
+import OrganizerEventList from "../../../components/market/organizer/OrganizerEventList"
 import OrganizerEventSearch from "../../../components/market/organizer/OrganizerEventSearch"
 import OrganizerInfoCard from "../../../components/market/organizer/OrganizerInfoCard"
 import BannerLayout from "../../../components/UI/layout/BannerLayout"
