+++ utils/generate/layer.ts
@@ -146,7 +146,7 @@ export function formatLayeredAssetMetadata(
   blobs: Blob[]
 ) {
   const regularTickets: TicketMetadataBlob[] = []
-  const vipTickets: TicketMetadataBlob[] = []
+  const vip: TicketMetadataBlob[] = []
 
   metadata.forEach((data, i) => {
     const ticket: TicketsMetadata = {
@@ -162,7 +162,7 @@ export function formatLayeredAssetMetadata(
     }
 
     if (data.hasVipAsset) {
-      vipTickets.push({
+      vip.push({
         metadata: ticket,
         blob: blobs[i]
       })
@@ -174,5 +174,5 @@ export function formatLayeredAssetMetadata(
     }
   })
 
-  return { regulars: regularTickets, vips: vipTickets }
+  return { regulars: regularTickets, vips: vip }
 }
